Address @pitrou code review comments
