+++ cpp/src/arrow/python/numpy-internal.h
@@ -170,239 +170,6 @@ inline bool PyIntScalar_Check(PyObject* obj) {
 
 }  // namespace internal
 
-// Form a consensus dtype to use for Arrow conversion for a collection of dtype
-// objects observed one at a time
-
-#define _NUMPY_UNIFY_NOOP(DTYPE) \
-  case NPY_##DTYPE:              \
-    return NOOP;
-
-#define _NUMPY_UNIFY_PROMOTE(DTYPE) \
-  case NPY_##DTYPE:                 \
-    return PROMOTE;
-
-class NumPyDtypeUnifier {
- public:
-  enum Action { NOOP, PROMOTE, INVALID };
-
-  NumPyDtypeUnifier() : current_type_num_(-1), current_dtype_(NULLPTR) {}
-
-  Status InvalidMix(int new_dtype) {
-    std::stringstream ss;
-    ss << "Cannot mix NumPy dtypes " << GetNumPyTypeName(current_type_num_) << " and "
-       << GetNumPyTypeName(new_dtype);
-    return Status::Invalid(ss.str());
-  }
-
-  int Observe_BOOL(PyArray_Descr* descr, int dtype) { return INVALID; }
-
-  int Observe_INT8(PyArray_Descr* descr, int dtype) {
-    switch (dtype) {
-      _NUMPY_UNIFY_PROMOTE(INT16);
-      _NUMPY_UNIFY_PROMOTE(INT32);
-      _NUMPY_UNIFY_PROMOTE(INT64);
-      _NUMPY_UNIFY_PROMOTE(FLOAT32);
-      _NUMPY_UNIFY_PROMOTE(FLOAT64);
-      default:
-        return INVALID;
-    }
-  }
-
-  int Observe_INT16(PyArray_Descr* descr, int dtype) {
-    switch (dtype) {
-      _NUMPY_UNIFY_NOOP(INT8);
-      _NUMPY_UNIFY_PROMOTE(INT32);
-      _NUMPY_UNIFY_PROMOTE(INT64);
-      _NUMPY_UNIFY_NOOP(UINT8);
-      _NUMPY_UNIFY_PROMOTE(FLOAT32);
-      _NUMPY_UNIFY_PROMOTE(FLOAT64);
-      default:
-        return INVALID;
-    }
-  }
-
-  int Observe_INT32(PyArray_Descr* descr, int dtype) {
-    switch (dtype) {
-      _NUMPY_UNIFY_NOOP(INT8);
-      _NUMPY_UNIFY_NOOP(INT16);
-      _NUMPY_UNIFY_PROMOTE(INT32);
-      _NUMPY_UNIFY_PROMOTE(INT64);
-      _NUMPY_UNIFY_NOOP(UINT8);
-      _NUMPY_UNIFY_NOOP(UINT16);
-      _NUMPY_UNIFY_PROMOTE(FLOAT32);
-      _NUMPY_UNIFY_PROMOTE(FLOAT64);
-      default:
-        return INVALID;
-    }
-  }
-
-  int Observe_INT64(PyArray_Descr* descr, int dtype) {
-    switch (dtype) {
-      _NUMPY_UNIFY_NOOP(INT8);
-      _NUMPY_UNIFY_NOOP(INT16);
-      _NUMPY_UNIFY_NOOP(INT32);
-      _NUMPY_UNIFY_NOOP(INT64);
-      _NUMPY_UNIFY_NOOP(UINT8);
-      _NUMPY_UNIFY_NOOP(UINT16);
-      _NUMPY_UNIFY_NOOP(UINT32);
-      _NUMPY_UNIFY_PROMOTE(FLOAT32);
-      _NUMPY_UNIFY_PROMOTE(FLOAT64);
-      default:
-        return INVALID;
-    }
-  }
-
-  int Observe_UINT8(PyArray_Descr* descr, int dtype) {
-    switch (dtype) {
-      _NUMPY_UNIFY_PROMOTE(UINT16);
-      _NUMPY_UNIFY_PROMOTE(UINT32);
-      _NUMPY_UNIFY_PROMOTE(UINT64);
-      _NUMPY_UNIFY_PROMOTE(FLOAT32);
-      _NUMPY_UNIFY_PROMOTE(FLOAT64);
-      default:
-        return INVALID;
-    }
-  }
-
-  int Observe_UINT16(PyArray_Descr* descr, int dtype) {
-    switch (dtype) {
-      _NUMPY_UNIFY_NOOP(UINT8);
-      _NUMPY_UNIFY_PROMOTE(UINT32);
-      _NUMPY_UNIFY_PROMOTE(UINT64);
-      _NUMPY_UNIFY_PROMOTE(FLOAT32);
-      _NUMPY_UNIFY_PROMOTE(FLOAT64);
-      default:
-        return INVALID;
-    }
-  }
-
-  int Observe_UINT32(PyArray_Descr* descr, int dtype) {
-    switch (dtype) {
-      _NUMPY_UNIFY_NOOP(UINT8);
-      _NUMPY_UNIFY_NOOP(UINT16);
-      _NUMPY_UNIFY_PROMOTE(UINT64);
-      _NUMPY_UNIFY_PROMOTE(FLOAT32);
-      _NUMPY_UNIFY_PROMOTE(FLOAT64);
-      default:
-        return INVALID;
-    }
-  }
-
-  int Observe_UINT64(PyArray_Descr* descr, int dtype) {
-    switch (dtype) {
-      _NUMPY_UNIFY_NOOP(UINT8);
-      _NUMPY_UNIFY_NOOP(UINT16);
-      _NUMPY_UNIFY_NOOP(UINT32);
-      _NUMPY_UNIFY_PROMOTE(FLOAT32);
-      _NUMPY_UNIFY_PROMOTE(FLOAT64);
-      default:
-        return INVALID;
-    }
-  }
-
-  int Observe_FLOAT16(PyArray_Descr* descr, int dtype) {
-    switch (dtype) {
-      _NUMPY_UNIFY_PROMOTE(FLOAT32);
-      _NUMPY_UNIFY_PROMOTE(FLOAT64);
-      default:
-        return INVALID;
-    }
-  }
-
-  int Observe_FLOAT32(PyArray_Descr* descr, int dtype) {
-    switch (dtype) {
-      _NUMPY_UNIFY_NOOP(INT8);
-      _NUMPY_UNIFY_NOOP(INT16);
-      _NUMPY_UNIFY_NOOP(INT32);
-      _NUMPY_UNIFY_NOOP(INT64);
-      _NUMPY_UNIFY_NOOP(UINT8);
-      _NUMPY_UNIFY_NOOP(UINT16);
-      _NUMPY_UNIFY_NOOP(UINT32);
-      _NUMPY_UNIFY_NOOP(UINT64);
-      _NUMPY_UNIFY_PROMOTE(FLOAT64);
-      default:
-        return INVALID;
-    }
-  }
-
-  int Observe_FLOAT64(PyArray_Descr* descr, int dtype) {
-    switch (dtype) {
-      _NUMPY_UNIFY_NOOP(INT8);
-      _NUMPY_UNIFY_NOOP(INT16);
-      _NUMPY_UNIFY_NOOP(INT32);
-      _NUMPY_UNIFY_NOOP(INT64);
-      _NUMPY_UNIFY_NOOP(UINT8);
-      _NUMPY_UNIFY_NOOP(UINT16);
-      _NUMPY_UNIFY_NOOP(UINT32);
-      _NUMPY_UNIFY_NOOP(UINT64);
-      default:
-        return INVALID;
-    }
-  }
-
-  int Observe_DATETIME(PyArray_Descr* dtype_obj) {
-    // TODO: check that units are all the same
-    // current_dtype_ = dtype_obj->type_num;
-    return NOOP;
-  }
-
-  Status Observe(PyArray_Descr* descr) {
-    const int dtype = fix_numpy_type_num(descr->type_num);
-
-    if (current_type_num_ == -1) {
-      current_dtype_ = descr;
-      current_type_num_ = dtype;
-      return Status::OK();
-    } else if (current_type_num_ == dtype) {
-      return Status::OK();
-    }
-
-#define OBSERVE_CASE(DTYPE)                 \
-  case NPY_##DTYPE:                         \
-    action = Observe_##DTYPE(descr, dtype); \
-    break;
-
-    int action = NOOP;
-    switch (current_type_num_) {
-      OBSERVE_CASE(BOOL);
-      OBSERVE_CASE(INT8);
-      OBSERVE_CASE(INT16);
-      OBSERVE_CASE(INT32);
-      OBSERVE_CASE(INT64);
-      OBSERVE_CASE(UINT8);
-      OBSERVE_CASE(UINT16);
-      OBSERVE_CASE(UINT32);
-      OBSERVE_CASE(UINT64);
-      OBSERVE_CASE(FLOAT16);
-      OBSERVE_CASE(FLOAT32);
-      OBSERVE_CASE(FLOAT64);
-      case NPY_DATETIME:
-        action = Observe_DATETIME(descr);
-        break;
-      default:
-        std::stringstream ss;
-        ss << "Unsupported numpy type " << GetNumPyTypeName(dtype) << std::endl;
-        return Status::NotImplemented(ss.str());
-    }
-
-    if (action == INVALID) {
-      return InvalidMix(dtype);
-    } else if (action == PROMOTE) {
-      current_type_num_ = dtype;
-      current_dtype_ = descr;
-    }
-    return Status::OK();
-  }
-
-  bool dtype_was_observed() const { return current_type_num_ != -1; }
-
-  PyArray_Descr* current_dtype() const { return current_dtype_; }
-
- private:
-  int current_type_num_;
-  PyArray_Descr* current_dtype_;
-};
-
 }  // namespace py
 }  // namespace arrow
