+++ cpp/src/arrow/python/numpy_convert.cc
@@ -276,78 +276,5 @@ Status TensorToNdarray(const std::shared_ptr<Tensor>& tensor, PyObject* base,
   return Status::OK();
 }
 
-Status UnboxIntegerAsInt64(PyObject* obj, int64_t* out) {
-  if (PyLong_Check(obj)) {
-    int overflow = 0;
-    *out = PyLong_AsLongLongAndOverflow(obj, &overflow);
-    if (overflow) {
-      return Status::Invalid("PyLong is too large to fit int64");
-    }
-#if PY_MAJOR_VERSION < 3
-  } else if (PyInt_Check(obj)) {
-    *out = static_cast<int64_t>(PyInt_AS_LONG(obj));
-#endif
-  } else if (PyArray_IsScalar(obj, UByte)) {
-    *out = reinterpret_cast<PyUByteScalarObject*>(obj)->obval;
-  } else if (PyArray_IsScalar(obj, Short)) {
-    *out = reinterpret_cast<PyShortScalarObject*>(obj)->obval;
-  } else if (PyArray_IsScalar(obj, UShort)) {
-    *out = reinterpret_cast<PyUShortScalarObject*>(obj)->obval;
-  } else if (PyArray_IsScalar(obj, Int)) {
-    *out = reinterpret_cast<PyIntScalarObject*>(obj)->obval;
-  } else if (PyArray_IsScalar(obj, UInt)) {
-    *out = reinterpret_cast<PyUIntScalarObject*>(obj)->obval;
-  } else if (PyArray_IsScalar(obj, Long)) {
-    *out = reinterpret_cast<PyLongScalarObject*>(obj)->obval;
-  } else if (PyArray_IsScalar(obj, ULong)) {
-    *out = reinterpret_cast<PyULongScalarObject*>(obj)->obval;
-  } else if (PyArray_IsScalar(obj, LongLong)) {
-    *out = reinterpret_cast<PyLongLongScalarObject*>(obj)->obval;
-  } else if (PyArray_IsScalar(obj, Int64)) {
-    *out = reinterpret_cast<PyInt64ScalarObject*>(obj)->obval;
-  } else if (PyArray_IsScalar(obj, ULongLong)) {
-    *out = reinterpret_cast<PyULongLongScalarObject*>(obj)->obval;
-  } else if (PyArray_IsScalar(obj, UInt64)) {
-    *out = reinterpret_cast<PyUInt64ScalarObject*>(obj)->obval;
-  } else {
-    return Status::Invalid("Integer scalar type not recognized");
-  }
-  return Status::OK();
-}
-
-Status IntegerScalarToDoubleSafe(PyObject* obj, double* out) {
-  int64_t value = 0;
-  RETURN_NOT_OK(UnboxIntegerAsInt64(obj, &value));
-
-  constexpr int64_t kDoubleMax = 1LL << 53;
-  constexpr int64_t kDoubleMin = -(1LL << 53);
-
-  if (value < kDoubleMin || value > kDoubleMax) {
-    std::stringstream ss;
-    ss << "Integer value " << value << " is outside of the range exactly"
-       << " representable by a IEEE 754 double precision value";
-    return Status::Invalid(ss.str());
-  }
-  *out = static_cast<double>(value);
-  return Status::OK();
-}
-
-Status IntegerScalarToFloat32Safe(PyObject* obj, float* out) {
-  int64_t value = 0;
-  RETURN_NOT_OK(UnboxIntegerAsInt64(obj, &value));
-
-  constexpr int64_t kFloatMax = 1LL << 24;
-  constexpr int64_t kFloatMin = -(1LL << 24);
-
-  if (value < kFloatMin || value > kFloatMax) {
-    std::stringstream ss;
-    ss << "Integer value " << value << " is outside of the range exactly"
-       << " representable by a IEEE 754 single precision value";
-    return Status::Invalid(ss.str());
-  }
-  *out = static_cast<float>(value);
-  return Status::OK();
-}
-
 }  // namespace py
 }  // namespace arrow
