+++ cpp/src/arrow/python/numpy_convert.h
@@ -68,9 +68,6 @@ ARROW_EXPORT Status NdarrayToTensor(MemoryPool* pool, PyObject* ao,
 ARROW_EXPORT Status TensorToNdarray(const std::shared_ptr<Tensor>& tensor, PyObject* base,
                                     PyObject** out);
 
-Status IntegerScalarToDoubleSafe(PyObject* obj, double* result);
-Status IntegerScalarToFloat32Safe(PyObject* obj, float* result);
-
 }  // namespace py
 }  // namespace arrow
